requirements.txt
ftfy==5.8
tqdm==4.49.0
torchvision==0.8.1
torch==1.7.0
regex==2020.11.13
streamlit==0.75.0
requests==2.25.0
Pillow==8.1.0
